requirements-dev.txt
# These are required for developing the package (running the tests, building
# the documentation) but not necessarily required for _using_ it.
python>=3.6.0
numpy
pip
sphinx
pytest
pytest-cov
flake8
lmfit
versioneer
pyqt
seabreeze
matplotlib
ipython-genutils==0.2.0
prompt-toolkit==1.0.15
sphinx-rtd-theme
super-state-machine
black

